[feat]添加获取个人中心页面信息的方法，完成了个人信息界面的构建
