+++ .idea/compiler.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="CompilerConfiguration">
-    <bytecodeTargetLevel target="14">
+    <bytecodeTargetLevel target="11">
       <module name="build_logic.convention" target="1.8" />
       <module name="build_logic.convention.main" target="1.8" />
       <module name="build_logic.convention.test" target="1.8" />
