+++ module_personal/build.gradle.kts
@@ -1,3 +1,7 @@
+import com.qxy.codeerror.convention.depend.*
+
 plugins {
     id("module-manager")
 }
+
+dependNetwork()
