+++ module_personal/src/main/java/com/qxy/module/personal/Constants.kt
@@ -1,0 +1,3 @@
+package com.qxy.module.personal
+
+const val KEY_PERSONAL_INFO = "personal_info_key"
